+++ src/main/kotlin/io/github/pandier/intellijdiscordrp/activity/ActivityFactory.kt
@@ -1,16 +1,21 @@
 package io.github.pandier.intellijdiscordrp.activity
 
 import io.github.pandier.intellijdiscordrp.settings.ImageSetting
+import io.github.pandier.intellijdiscordrp.settings.LogoStyleSetting
 import io.github.vyfor.kpresence.rpc.Activity
 import io.github.vyfor.kpresence.rpc.activity
 
-private fun ImageSetting.getIcon(context: ActivityContext) = when (this) {
-    ImageSetting.APPLICATION -> currentActivityApplicationType.icon
+private fun ImageSetting.getIcon(context: ActivityContext, logoStyle: LogoStyleSetting) = when (this) {
+    ImageSetting.APPLICATION -> when (logoStyle) {
+        LogoStyleSetting.MODERN -> currentActivityApplicationType.modernIcon
+        LogoStyleSetting.CLASSIC -> currentActivityApplicationType.classicIcon
+    }
     ImageSetting.FILE -> context.file?.type?.icon
 }
 
 class ActivityFactory(
     private val displayMode: ActivityDisplayMode,
+    private val logoStyle: LogoStyleSetting,
     private val details: String,
     private val state: String,
     private val largeImage: ImageSetting?,
@@ -20,18 +25,18 @@ class ActivityFactory(
     private val timestampEnabled: Boolean,
 ) {
     fun create(context: ActivityContext): Activity = activity {
-        details = this@ActivityFactory.details.ifEmpty { null }?.let { displayMode.format(it, context) }
-        state = this@ActivityFactory.state.ifEmpty { null }?.let { displayMode.format(it, context) }
+        details = this@ActivityFactory.details.ifEmpty { null }?.let { displayMode.format(it, context).padEnd(2) }
+        state = this@ActivityFactory.state.ifEmpty { null }?.let { displayMode.format(it, context).padEnd(2) }
 
         assets {
             if (this@ActivityFactory.largeImage != null) {
-                largeImage = this@ActivityFactory.largeImage.getIcon(context)
-                largeText = displayMode.format(this@ActivityFactory.largeImageText, context)
+                largeImage = this@ActivityFactory.largeImage.getIcon(context, logoStyle)
+                largeText = displayMode.format(this@ActivityFactory.largeImageText, context).padEnd(2)
             }
 
             if (this@ActivityFactory.smallImage != null) {
-                smallImage = this@ActivityFactory.smallImage.getIcon(context)
-                smallText = displayMode.format(this@ActivityFactory.smallImageText, context)
+                smallImage = this@ActivityFactory.smallImage.getIcon(context, logoStyle)
+                smallText = displayMode.format(this@ActivityFactory.smallImageText, context).padEnd(2)
             }
         }